refactors

+++ bootstrap.sh
@@ -26,7 +26,9 @@ pacman-key --populate archlinuxarm
 #timedatectl set-ntp true
 
 # Enable network connection
-mv /etc/resolv.conf /etc/resolv.conf.bk
+if [[ -f /etc/resolv.conf ]]; then
+  mv /etc/resolv.conf /etc/resolv.conf.bk;
+fi
 echo 'nameserver 8.8.8.8' > /etc/resolv.conf
 pacman -Sy --noconfirm --needed
 
@@ -106,4 +108,7 @@ chmod 700 "/home/${username}/.ssh"
 chmod 600 "/home/${username}/.ssh/authorized_keys"
 
 # restore original resolve.conf
-mv /etc/resolv.conf.bk /etc/resolv.conf
+if [[ -f /etc/resolv.conf.bk ]]; then
+  mv /etc/resolv.conf.bk /etc/resolv.conf;
+fi
+
